--- a/.pnp.cjs
+++ b/.pnp.cjs
@@ -8467,15 +8467,6 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["@types/js-yaml", [\
-      ["npm:3.12.5", {\
-        "packageLocation": "./.yarn/cache/@types-js-yaml-npm-3.12.5-9b6b585722-15d44f172e.zip/node_modules/@types/js-yaml/",\
-        "packageDependencies": [\
-          ["@types/js-yaml", "npm:3.12.5"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
     ["@types/json-patch", [\
       ["npm:0.0.30", {\
         "packageLocation": "./.yarn/cache/@types-json-patch-npm-0.0.30-1590e68edc-44e36c8c27.zip/node_modules/@types/json-patch/",\
@@ -11206,10 +11197,9 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./packages/yarnpkg-parsers/",\
         "packageDependencies": [\
           ["@yarnpkg/parsers", "workspace:packages/yarnpkg-parsers"],\
-          ["@types/js-yaml", "npm:3.12.5"],\
-          ["js-yaml", "npm:3.14.1"],\
           ["pegjs", "npm:0.10.0"],\
-          ["tslib", "npm:2.4.0"]\
+          ["tslib", "npm:2.4.0"],\
+          ["wasm-pack", "npm:0.10.3"]\
         ],\
         "linkType": "SOFT"\
       }]\
@@ -22199,6 +22189,18 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/binary-extensions-npm-2.0.0-8343f65d59-d9aeff7603.zip/node_modules/binary-extensions/",\
         "packageDependencies": [\
           ["binary-extensions", "npm:2.0.0"]\
+        ],\
+        "linkType": "HARD"\
+      }]\
+    ]],\
+    ["binary-install", [\
+      ["npm:0.1.1", {\
+        "packageLocation": "./.yarn/unplugged/binary-install-npm-0.1.1-7d7f3ccef2/node_modules/binary-install/",\
+        "packageDependencies": [\
+          ["binary-install", "npm:0.1.1"],\
+          ["axios", "npm:0.21.1"],\
+          ["rimraf", "npm:3.0.2"],\
+          ["tar", "npm:6.1.11"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -42346,6 +42348,19 @@ const RAW_RUNTIME_STATE =
           ["yallist", "npm:4.0.0"]\
         ],\
         "linkType": "HARD"\
+      }],\
+      ["npm:6.1.11", {\
+        "packageLocation": "./.yarn/cache/tar-npm-6.1.11-e6ac3cba9c-5499de6e19.zip/node_modules/tar/",\
+        "packageDependencies": [\
+          ["tar", "npm:6.1.11"],\
+          ["chownr", "npm:2.0.0"],\
+          ["fs-minipass", "npm:2.0.0"],\
+          ["minipass", "npm:3.1.3"],\
+          ["minizlib", "npm:2.1.2"],\
+          ["mkdirp", "npm:1.0.4"],\
+          ["yallist", "npm:4.0.0"]\
+        ],\
+        "linkType": "HARD"\
       }]\
     ]],\
     ["tar-fs", [\
@@ -44480,6 +44495,16 @@ const RAW_RUNTIME_STATE =
         "packageDependencies": [\
           ["warning", "npm:4.0.3"],\
           ["loose-envify", "npm:1.4.0"]\
+        ],\
+        "linkType": "HARD"\
+      }]\
+    ]],\
+    ["wasm-pack", [\
+      ["npm:0.10.3", {\
+        "packageLocation": "./.yarn/unplugged/wasm-pack-npm-0.10.3-fba77b1879/node_modules/wasm-pack/",\
+        "packageDependencies": [\
+          ["wasm-pack", "npm:0.10.3"],\
+          ["binary-install", "npm:0.1.1"]\
         ],\
         "linkType": "HARD"\
       }]\