--- a/.yarnrc.yml
+++ b/.yarnrc.yml
@@ -10,9 +10,15 @@ pnpEnableEsmLoader: false
 immutablePatterns:
   - .pnp.*
 
+# TODO: Revert these to flow sequences.
 supportedArchitectures:
-  os: [darwin, linux, win32]
-  cpu: [x64, arm64]
+  os:
+    - darwin
+    - linux
+    - win32
+  cpu:
+    - x64
+    - arm64
 
 initScope: yarnpkg