--- a/packages/yarnpkg-parsers/package.json
+++ b/packages/yarnpkg-parsers/package.json
@@ -4,17 +4,16 @@
   "license": "BSD-2-Clause",
   "main": "./sources/index.ts",
   "dependencies": {
-    "js-yaml": "^3.10.0",
     "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@types/js-yaml": "^3.11.1",
-    "pegjs": "^0.10.0"
+    "pegjs": "^0.10.0",
+    "wasm-pack": "^0.10.3"
   },
   "scripts": {
     "grammar:shell": "run pegjs -o sources/grammars/shell.js sources/grammars/shell.pegjs",
     "grammar:resolution": "run pegjs -o sources/grammars/resolution.js sources/grammars/resolution.pegjs",
-    "grammar:syml": "run pegjs -o sources/grammars/syml.js sources/grammars/syml.pegjs",
+    "grammar:syml": "cd ./artifacts && ./build.sh",
     "grammar:all": "run grammar:shell & run grammar:resolution & run grammar:syml &",
     "postpack": "rm -rf lib",
     "prepack": "run build:compile \"$(pwd)\"",