--- a/packages/yarnpkg-parsers/sources/grammars/syml.d.ts
+++ b/packages/yarnpkg-parsers/sources/grammars/syml.d.ts
@@ -1,1 +1,7 @@
-export declare const parse: (code: string) => {[key: string]: any};
+/* tslint:disable */
+/* eslint-disable */
+/**
+* @param {string} input
+* @returns {any}
+*/
+export function parse(input: string): any;