--- a/rust-toolchain.toml
+++ b/rust-toolchain.toml
@@ -1,0 +1,6 @@
+[toolchain]
+channel = "nightly-2022-08-31"
+profile = "default"
+# Required for `build-std`
+components = ["rust-src"]
+targets = ["wasm32-unknown-unknown"]