--- a/rustfmt.toml
+++ b/rustfmt.toml
@@ -1,0 +1,2 @@
+tab_spaces = 2
+max_width = 100