@paul-soporan 2022-09-03
A few more notes:
- compiling the parser to a native node addon using `napi` makes it 40% faster than the WASM version; unfortunately we can't use it due to our portability requirements
- generally nom parsers that work on `&[u8]` are faster than ones that work on `&str`, I need to test whether it would improve things
- it looks like `wasm-bindgen` doesn't support BE architectures at all - all of the glue code assumes LE
- `rustc` supports [Profile Guided Optimizations (PGOs)](https://doc.rust-lang.org/rustc/profile-guided-optimization.html) natively; not sure whether they work for WASM targets too, but it might be interesting to explore