@paul-soporan 2022-09-03
Update: I changed the parser to work on `&[u8]` instead of `&str` and it's now 20% faster than before.

For a single lockfile parse it's about 30% faster than `js-yaml` while for 10 lockfile parses it's about 6% slower.

The bundle size also went down from `2.72 MB` to `2.71 MB`.